# Optimize CKKS square