@WeiDaiWD 2021-06-15
*Testing with clang-10:*
This PR is much faster than original SEAL, the tiled version is not (much) faster than this PR.

*Testing with gcc-9:*
The tiled version is faster than this PR which is much faster than original SEAL, except for 1024.
For 1024, the original SEAL is as fast as clang-10, costing 22 us, while both new versions with gcc-9 take 60+ us which is slower than 2048. I disassembled `Evaluator::ckks_square` in the original SEAL and this PR, and they are almost identical. The issue is likely caused by poor/weird performance of `SEAL_ITERATE` with GCC, which is called in `dyadic_product`.

I'll merge this PR. For gcc and 1024 case, I'll look into fixing that in a future release. I don't think this affects too many users. The speedup we get from this PR is more valuable. Thank you so much for this!